add layout and components for lesson 3
