+++ src/components/HelloWorld.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <p>Hello Main!</p>
+  </div>
+</template>
