+++ src/layouts/components/Footer.vue
