+++ src/views/Home.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="home">
+    <HelloWorld msg="Welcome to Your Vue.js App"/>
+  </div>
+</template>
+
+<script>
+// @ is an alias to /src
+import HelloWorld from '@/components/HelloWorld.vue';
+
+export default {
+  name: 'Home',
+  components: {
+    HelloWorld,
+  },
+};
+</script>
