+++ src/views/shop/components/how-to-balls.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <p>Hello Balls!</p>
+  </div>
+</template>
